+++ src/Education.Client/Gateways/History/UserTopicProgress.cs
@@ -2,4 +2,4 @@ using System;
 
 namespace Education.Client.Gateways.History;
 
-public sealed record UserTopicProgress(uint PassedTests, TimeSpan TimeSpent);
+public sealed record UserTopicProgress(ulong PassedTests, TimeSpan TimeSpent);
